Поготовлена к внеднерию в основную ветку. Админ панель и управление ролями.
