+++ app/dashboard/users/admin/page.tsx
@@ -1,0 +1,5 @@
+import { AdminPage } from "@/modules/users/admin"
+
+export default function AdminRoute() {
+  return <AdminPage />
+}
