+++ docs/admin_refactor_plan.md
@@ -1,0 +1,191 @@
+# План рефакторинга админ-панели
+
+## Цель рефакторинга
+Перенести функционал админ-панели из `app/dashboard/admin` в модульную структуру `modules/users/admin` с новым роутом `/dashboard/users/admin`.
+
+## Текущая структура (что есть сейчас)
+```
+app/dashboard/admin/
+├── page.tsx (1.4KB, 50 lines) - главная страница админки
+├── AdminPanel.tsx (1.6KB, 38 lines) - основной компонент панели
+├── RolesTab.tsx (27KB, 739 lines) - управление ролями
+├── README.md (5.1KB, 101 lines) - документация
+├── components/ - папка с компонентами (будет удалена)
+├── CategoriesTab.tsx (6.3KB, 183 lines) - управление категориями
+├── DepartmentsTab.tsx (7.1KB, 202 lines) - управление департаментами  
+├── PositionsTab.tsx (6.3KB, 183 lines) - управление позициями
+└── TeamsTab.tsx (9.5KB, 274 lines) - управление командами
+```
+
+## Целевая структура (что будет после рефакторинга)
+
+### 1. Новая модульная структура
+```
+modules/users/admin/
+├── AdminPage.tsx - главная страница админки (перенос из page.tsx)
+├── AdminPanel.tsx - основной компонент панели
+├── components/
+│   ├── RolesTab.tsx - управление ролями
+│   ├── CategoriesTab.tsx - управление категориями
+│   ├── DepartmentsTab.tsx - управление департаментами
+│   ├── PositionsTab.tsx - управление позициями
+│   ├── TeamsTab.tsx - управление командами
+│   └── index.ts - экспорты компонентов
+├── hooks/
+│   └── useAdminPermissions.ts - хук для проверки админских прав
+├── types/
+│   └── admin.ts - типы для админ-панели
+├── utils/
+│   └── adminHelpers.ts - вспомогательные функции
+├── README.md - документация модуля
+└── index.ts - главный экспорт модуля
+```
+
+### 2. Новый роут
+```
+app/dashboard/users/admin/
+└── page.tsx - новая страница, импортирующая AdminPage из модуля
+```
+
+### 3. Удаление старой структуры
+```
+app/dashboard/admin/ - УДАЛИТЬ ПОЛНОСТЬЮ
+```
+
+## Детальный план выполнения
+
+### Этап 1: Создание новой структуры папок ✅
+1. **Создать модульную структуру**
+   - `modules/users/admin/`
+   - `modules/users/admin/components/`
+   - `modules/users/admin/hooks/`
+   - `modules/users/admin/types/`
+   - `modules/users/admin/utils/`
+
+2. **Создать новый роут**
+   - `app/dashboard/users/admin/`
+
+### Этап 2: Перенос компонентов ✅
+1. **Перенести основные файлы**
+   - `page.tsx` → `modules/users/admin/AdminPage.tsx`
+   - `AdminPanel.tsx` → `modules/users/admin/AdminPanel.tsx`
+   - `README.md` → `modules/users/admin/README.md`
+
+2. **Перенести табы в components/**
+   - `RolesTab.tsx` → `modules/users/admin/components/RolesTab.tsx`
+   - `CategoriesTab.tsx` → `modules/users/admin/components/CategoriesTab.tsx`
+   - `DepartmentsTab.tsx` → `modules/users/admin/components/DepartmentsTab.tsx`
+   - `PositionsTab.tsx` → `modules/users/admin/components/PositionsTab.tsx`
+   - `TeamsTab.tsx` → `modules/users/admin/components/TeamsTab.tsx`
+
+### Этап 3: Создание вспомогательных файлов ✅
+1. **Создать экспорты**
+   - `modules/users/admin/components/index.ts`
+   - `modules/users/admin/index.ts`
+
+2. **Создать вспомогательные файлы**
+   - `modules/users/admin/types/admin.ts`
+   - `modules/users/admin/hooks/useAdminPermissions.ts`
+   - `modules/users/admin/utils/adminHelpers.ts`
+
+### Этап 4: Исправление импортов ✅
+1. **Обновить все импорты в перенесенных файлах**
+2. **Проверить работоспособность компонентов**
+
+### Этап 5: Создание нового роута ✅
+1. **Создать app/dashboard/users/admin/page.tsx**
+   - Импортировать и рендерить AdminPage из модуля
+
+### Этап 6: Удаление старой структуры ✅
+1. **Удалить app/dashboard/admin/ полностью**
+
+### Этап 7: Обновление навигации ✅
+1. **Обновить ссылки в сайдбаре и других местах**
+   - Изменить `/dashboard/admin` на `/dashboard/users/admin`
+
+## Согласованные решения
+
+1. ✅ **Роутинг**: Создаем новый роут `/dashboard/users/admin`
+2. ✅ **Store**: НЕ создаем отдельный store
+3. ✅ **Папка components/**: Удаляем существующую папку
+4. ✅ **Обратная совместимость**: НЕ нужна, только новый функционал
+5. ✅ **Тестирование**: НЕ создаем тесты
+
+## Принципы выполнения
+
+1. **Строгий перенос**: Переносим только существующий функционал без изменений
+2. **Работоспособность**: Все должно работать как раньше
+3. **Чистота**: Удаляем старые файлы полностью
+4. **Новая архитектура**: Следуем модульной структуре
+
+## Порядок выполнения
+
+1. ✅ Согласование плана
+2. ✅ Создание структуры папок
+3. ✅ Перенос основных компонентов
+4. ✅ Создание вспомогательных файлов
+5. ✅ Исправление импортов
+6. ✅ Создание нового роута
+7. ✅ Удаление старой структуры
+8. ✅ Обновление навигации
+9. ✅ Финальная проверка
+
+---
+
+**✅ РЕФАКТОРИНГ ЗАВЕРШЕН УСПЕШНО!**
+
+## Результат рефакторинга
+
+### ✅ Что было выполнено:
+
+1. **Создана новая модульная структура** `modules/users/admin/`
+2. **Перенесены все компоненты** из старой папки в новую структуру
+3. **Создан новый роут** `/dashboard/users/admin`
+4. **Удалена старая структура** `app/dashboard/admin/`
+5. **Обновлены все импорты** в существующих файлах
+6. **Созданы недостающие компоненты**: EntityModal, DeleteConfirmModal, LoadingState, EmptyState
+7. **Добавлены вспомогательные файлы**: типы, хуки, утилиты
+8. **Проект успешно собирается** без ошибок
+
+### 📁 Финальная структура:
+
+```
+modules/users/admin/
+├── AdminPage.tsx - главная страница админки
+├── AdminPanel.tsx - основной компонент панели  
+├── components/
+│   ├── RolesTab.tsx - управление ролями
+│   ├── CategoriesTab.tsx - управление категориями
+│   ├── DepartmentsTab.tsx - управление департаментами
+│   ├── PositionsTab.tsx - управление позициями
+│   ├── TeamsTab.tsx - управление командами
+│   ├── EntityModal.tsx - модальное окно создания/редактирования
+│   ├── DeleteConfirmModal.tsx - модальное окно подтверждения удаления
+│   ├── LoadingState.tsx - состояние загрузки
+│   ├── EmptyState.tsx - пустое состояние
+│   └── index.ts - экспорты компонентов
+├── hooks/
+│   └── useAdminPermissions.ts - хук для проверки админских прав
+├── types/
+│   └── admin.ts - типы для админ-панели
+├── utils/
+│   └── adminHelpers.ts - вспомогательные функции
+├── README.md - документация модуля
+└── index.ts - главный экспорт модуля
+
+app/dashboard/users/admin/
+└── page.tsx - новая страница роута
+```
+
+### 🔗 Обновленная навигация:
+
+- Админ-панель доступна по адресу: `/dashboard/users/admin`
+- Интегрирована в существующую страницу пользователей через вкладку "Администратор"
+- Все импорты обновлены на новые пути
+
+### ✅ Проверка работоспособности:
+
+- ✅ Проект собирается без ошибок
+- ✅ Все компоненты имеют правильные импорты
+- ✅ Новый роут создан и работает
+- ✅ Старая структура полностью удалена
