+++ modules/users/admin/components/LoadingState.tsx
@@ -1,0 +1,23 @@
+import { Skeleton } from "@/components/ui/skeleton"
+import { TableRow, TableCell } from "@/components/ui/table"
+
+interface LoadingStateProps {
+  columnCount: number
+  rowCount?: number
+}
+
+export default function LoadingState({ columnCount, rowCount = 5 }: LoadingStateProps) {
+  return (
+    <>
+      {Array.from({ length: rowCount }).map((_, index) => (
+        <TableRow key={index}>
+          {Array.from({ length: columnCount }).map((_, cellIndex) => (
+            <TableCell key={cellIndex}>
+              <Skeleton className="h-4 w-full" />
+            </TableCell>
+          ))}
+        </TableRow>
+      ))}
+    </>
+  )
+}
