+++ modules/users/admin/components/index.ts
@@ -1,0 +1,9 @@
+export { default as RolesTab } from './RolesTab'
+export { default as CategoriesTab } from './CategoriesTab'
+export { default as DepartmentsTab } from './DepartmentsTab'
+export { default as PositionsTab } from './PositionsTab'
+export { default as TeamsTab } from './TeamsTab'
+export { default as EntityModal } from './EntityModal'
+export { default as DeleteConfirmModal } from './DeleteConfirmModal'
+export { default as LoadingState } from './LoadingState'
+export { default as EmptyState } from './EmptyState'
