+++ modules/users/admin/index.ts
@@ -1,0 +1,3 @@
+export { default as AdminPage } from './AdminPage'
+export { default as AdminPanel } from './AdminPanel'
+export * from './components'
