+++ modules/users/pages/users-page.tsx
@@ -10,7 +10,7 @@ import { getUsers } from "@/services/org-data-service"
 import { useState, useEffect } from "react"
 import type { User } from "@/types/db"
 import { PaymentAccessCheck } from "../components/payment-access-check"
-import AdminPanel from "@/app/dashboard/admin/AdminPanel"
+import { AdminPanel } from "@/modules/users/admin"
 import { useUserStore } from "@/stores/useUserStore"
 import { AdminAccessCheck } from "../components/admin-access-check"
 import { useSearchParams, useRouter } from "next/navigation"
